+++ gen/com/yanncebron/m68kplugin/lang/M68kDirectivesParser.java
@@ -1113,48 +1113,53 @@ public class M68kDirectivesParser {
   }
 
   /* ********************************************************** */
-  // MACRO_CALL_ID data_size_all? macro_call_parameter? (COMMA macro_call_parameter)*
+  // MACRO_CALL_ID
+  //                          <<enterMacroCall>>
+  //                          data_size_all? macro_call_parameter? (COMMA macro_call_parameter)*
+  //                          <<exitMacroCall>>
   public static boolean macro_call_directive(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "macro_call_directive")) return false;
     if (!nextTokenIs(b, "<directive>", MACRO_CALL_ID)) return false;
     boolean r;
     Marker m = enter_section_(b, l, _NONE_, MACRO_CALL_DIRECTIVE, "<directive>");
     r = consumeToken(b, MACRO_CALL_ID);
-    r = r && macro_call_directive_1(b, l + 1);
+    r = r && enterMacroCall(b, l + 1);
     r = r && macro_call_directive_2(b, l + 1);
     r = r && macro_call_directive_3(b, l + 1);
+    r = r && macro_call_directive_4(b, l + 1);
+    r = r && exitMacroCall(b, l + 1);
     exit_section_(b, l, m, r, false, null);
     return r;
   }
 
   // data_size_all?
-  private static boolean macro_call_directive_1(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "macro_call_directive_1")) return false;
+  private static boolean macro_call_directive_2(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "macro_call_directive_2")) return false;
     data_size_all(b, l + 1);
     return true;
   }
 
   // macro_call_parameter?
-  private static boolean macro_call_directive_2(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "macro_call_directive_2")) return false;
+  private static boolean macro_call_directive_3(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "macro_call_directive_3")) return false;
     macro_call_parameter(b, l + 1);
     return true;
   }
 
   // (COMMA macro_call_parameter)*
-  private static boolean macro_call_directive_3(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "macro_call_directive_3")) return false;
+  private static boolean macro_call_directive_4(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "macro_call_directive_4")) return false;
     while (true) {
       int c = current_position_(b);
-      if (!macro_call_directive_3_0(b, l + 1)) break;
-      if (!empty_element_parsed_guard_(b, "macro_call_directive_3", c)) break;
+      if (!macro_call_directive_4_0(b, l + 1)) break;
+      if (!empty_element_parsed_guard_(b, "macro_call_directive_4", c)) break;
     }
     return true;
   }
 
   // COMMA macro_call_parameter
-  private static boolean macro_call_directive_3_0(PsiBuilder b, int l) {
-    if (!recursion_guard_(b, l, "macro_call_directive_3_0")) return false;
+  private static boolean macro_call_directive_4_0(PsiBuilder b, int l) {
+    if (!recursion_guard_(b, l, "macro_call_directive_4_0")) return false;
     boolean r;
     Marker m = enter_section_(b);
     r = consumeToken(b, COMMA);
@@ -1164,12 +1169,13 @@ public class M68kDirectivesParser {
   }
 
   /* ********************************************************** */
-  // adm_group_all | adm_sr | adm_ccr | adm_usp | adm_group_ctrl_registers
+  // adm_register_list | adm_group_all | adm_sr | adm_ccr | adm_usp | adm_group_ctrl_registers
   public static boolean macro_call_parameter(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "macro_call_parameter")) return false;
     boolean r;
     Marker m = enter_section_(b, l, _NONE_, MACRO_CALL_PARAMETER, "<macro parameter>");
-    r = adm_group_all(b, l + 1);
+    r = adm_register_list(b, l + 1);
+    if (!r) r = adm_group_all(b, l + 1);
     if (!r) r = adm_sr(b, l + 1);
     if (!r) r = adm_ccr(b, l + 1);
     if (!r) r = adm_usp(b, l + 1);